Add files via upload
FINGERPRINT BASED VOTING SYSTEM
The main objective of democracy is "vote" by which the people can elect the candidates for forming an efficient
government to satisfy their needs and requests such that their standard living can be improved. In developing
countries like "INDIA", the election commission follows manual voting mechanism which is done by the electronic
voting machine. But instead of this, The poll rate of India has only increased by 4 percent from 1952 to 2014. So a
machine was required to automate the process and can be avail machine is placed in the poll booth center and is
monitored by higher officials, due to some illegal activities the polling center are misused and people's right to has
been denied. This seldom occurs in rural areas as well as in urban cities because the educated people are not
interested in casting their votes to candidates who represent their respective areas.
To ensure 100% voting, automation came into play. But this automated system has been approved only on some
developed countries since security have not been ensured to a large extent. The poll percentage of India have never
exceeded 67% till date. The main problem is people either do not leave in the area where they are registered as a
voter or they do not go to the poll center because of any other reason. The queue at the poll center is also one of the
major reason for less poll percentage. Some people cast their vote, and approximate 2% of the vote each time
become invalid, due to any reason. To overcome all the drawbacks of traditional methods of voting, We have come
up with an innovating solution to solve all the above-discussed problems. Fingerprint Based Voting Project is an
application where the user is recognized by his finger pattern or IRIS (or any other biometrics in the future). Since the
finger pattern of each human being is different, the voter can be easily authenticated. The system allows the voter to
vote through his fingerprint. The fingerprint is used to uniquely identify the user. The fingerprint minutiae features are
different for each human being. Fingerprint is used as an authentication of the voters. A voter can vote the candidate
only once, the system will not allow the candidate to vote for the second time.
Admin will register the voter's name by verifying voter. Admin will authenticate the user by verifying the user’s identity
proof and then admin will register the voter. Once the user has got the user id and password from the admin the user
can log in and vote for the candidate who is nominated. The system will allow the user to vote for only one candidate.
The system will allow the user to vote for one time for a particular election. Admin can add any number of candidates
when the new election will be announced. Admin can view the election result by using the election id. Even user can
view the election result. We with the help of this system aims to achieve a target of 80+ percent . We have designed
this system in such a way that is affordable with minimum hardware requirement and the software will be provided by
EC free of cost.